Update styles.css
Added a border to the article section to help make it easier to distinguish each article from one another. Worked on the spacing of each h2 tag within the article so it is not too far from the p tag. Worked on the placement of the time stamp of the article. Liking the way the page is displaying. will continue to mess with the styling as I learn more.