Update styles.css
Messing around with the navigation bar to make it a side bar that is fixed on the screen. Got it to the left hand side of the screen. Just need to work on getting the rest of the page to display around it. These changes are mostly to the nav styles